refactor: rename Windows projects
This renames Windows projects adding postfix -win.

+++ client-win/cli/RequestSelector.cpp
@@ -10,13 +10,13 @@ nlohmann::json cli::RequestSelector::select()
 	while (true)
 	{
 		std::cout << "Please enter request ID:" << '\n';
-		for (int i = 0; i < validRequests.size(); i++)
+		for (size_t i = 0; i < validRequests.size(); i++)
 		{
 			std::cout << i << " - " << validRequests.at(i) << '\n';
 		}
 		std::cout << std::endl;
 
-		int requestId{};
+		size_t requestId{};
 		if (!(std::cin >> requestId) || requestId < 0 || requestId >= validRequests.size())
 		{
 			std::cout << "The entered value is not a valid request ID." << std::endl;
